requirements.txt
# Core ML & Data
pandas>=1.3.0
numpy>=1.21.0
torch --index-url https://download.pytorch.org/whl/cpu
scikit-learn>=1.0.0
wandb>=0.15.0
requests

# Web API
fastapi
uvicorn[standard]
pydantic
gunicorn

# Plotting (for notebooks/analysis)
matplotlib

# Utilities
python-dotenv>=0.19.0
streamlit

# Jupyter
ipykernel
notebook
torchvision==0.16.2 
torchaudio==2.1.2

# Data
datasets
pyarrow
fastparquet>=0.8.0
psutil>=5.9.0  # For memory monitoring

chromadb
sentence-transformers

python-multipart
